updating to include a revisions table
