--- alembic/versions/337ed7fb992a_initial_schema.py
@@ -1,50 +1,0 @@
-"""Initial schema
-
-Revision ID: 337ed7fb992a
-Revises: 
-Create Date: 2025-07-05 11:52:31.790430
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '337ed7fb992a'
-down_revision: Union[str, Sequence[str], None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('links',
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('slug', sa.String(), nullable=False),
-    sa.Column('destination_url', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('slug')
-    )
-    op.create_table('clicks',
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('link_id', sa.UUID(), nullable=False),
-    sa.Column('timestamp', sa.DateTime(), nullable=True),
-    sa.Column('ip_address', sa.String(), nullable=True),
-    sa.Column('referrer', sa.String(), nullable=True),
-    sa.Column('user_agent', sa.String(), nullable=True),
-    sa.ForeignKeyConstraint(['link_id'], ['links.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('clicks')
-    op.drop_table('links')
-    # ### end Alembic commands ###
