+++ app/models.py
@@ -1,15 +1,30 @@
-from sqlalchemy import Column, String, DateTime, ForeignKey
-from sqlalchemy.dialects.postgresql import UUID
-from datetime import datetime
 import uuid
-from .db import Base
+from datetime import datetime
+from sqlalchemy import Column, String, Text, Integer, ForeignKey, DateTime
+from sqlalchemy.dialects.postgresql import UUID
+from sqlalchemy.orm import relationship
+from app.db import Base
 
 class Link(Base):
     __tablename__ = "links"
+
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
-    slug = Column(String, unique=True, nullable=False)
-    destination_url = Column(String, nullable=False)
+    slug = Column(String(255), unique=True, nullable=False)
     created_at = Column(DateTime, default=datetime.utcnow)
+
+    versions = relationship("LinkVersion", back_populates="link", cascade="all, delete-orphan", order_by="LinkVersion.version")
+
+
+class LinkVersion(Base):
+    __tablename__ = "link_versions"
+
+    id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
+    link_id = Column(UUID(as_uuid=True), ForeignKey("links.id", ondelete="CASCADE"), nullable=False)
+    version = Column(Integer, nullable=False)
+    destination_url = Column(Text, nullable=False)
+    created_at = Column(DateTime, default=datetime.utcnow)
+
+    link = relationship("Link", back_populates="versions")
 
 class Click(Base):
     __tablename__ = "clicks"
