+++ cej-deployment.yaml
@@ -1,0 +1,56 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: cej-deployment
+spec:
+  replicas: 2
+  selector:
+    matchLabels:
+      app: cej
+  template:
+    metadata:
+      labels:
+        app: cej
+    spec:
+      containers:
+      - name: cej
+        image: docker.io/younga8699/cej:latest
+        imagePullPolicy: Always
+        ports:
+        - containerPort: 8000
+        env:
+          - name: POSTGRES_USER
+            valueFrom:
+              secretKeyRef:
+                name: cej-secrets
+                key: POSTGRES_USER
+          - name: POSTGRES_PASSWORD
+            valueFrom:
+              secretKeyRef:
+                name: cej-secrets
+                key: POSTGRES_PASSWORD
+          - name: POSTGRES_DB
+            valueFrom:
+              secretKeyRef:
+                name: cej-secrets
+                key: POSTGRES_DB
+          - name: POSTGRES_HOST
+            valueFrom:
+              secretKeyRef:
+                name: cej-secrets
+                key: POSTGRES_HOST
+          - name: POSTGRES_PORT
+            value: "5432"
+---
+apiVersion: v1
+kind: Service
+metadata:
+  name: cej-service
+spec:
+  selector:
+    app: cej
+  ports:
+    - protocol: TCP
+      port: 80
+      targetPort: 8000
+  type: LoadBalancer
